ttools: protect KNN Plotter from unhappy combiner options
The KnnKernelDensityPlotter has not been fixed to work with
different aggregation modes.  I think it could be done, but
it seems rather special interest, don't bother until somebody
actually asks for it.

In the mean time, if it is given (via the ConfigMap) a non-SUM
combiner, it would do the wrong thing.  So write a unit test and
insert a runtime assertion to check that doesn't happen.