topcat: fix unselected rows regression bug
Fix a bug related to creating row subsets from selected rows
in the table viewer window.  When the subset is defined from
unselected rows, it should include only currently visible
unselected rows (ones in the current subset) but it was including
all unselected rows.

This was a regression bug introduced when adding support for
enormous JTables in 2014.  Reported by Ming Yang.

Also fixed a couple of minor mistakes in action names and tooltips.