ttools: prepare for density-like Combiners
The Combiner contract now requires that a bin factor is applied to
the aggregated bin values that are produced.  This can take account
of bin sizes so that density-like quantities can be calculated.

There are several changes to the code, but there *should* be no
change to behaviour at this commit.