ttools: remove configurable plotter normalisation keys
Remove the ConfigKey<Normalisation> parameter from the constructors
of the KDE, KNN, Histogram and Gaussian plotters.
They now use a fixed static config key to acquire normalisation.

This configurability was in there so that these plotters could
behave differently for Time and Plane plots.
This is still desirable, but doing it via the Normalisation parameter
is no longer the way I want to do it.  So remove this code,
in preparation for adding a different way to configure
time-unit-sensitive bar height scaling in the future.