ttools: do density-like scaling in Normalisation class
Combiner.Type scaling for histogram-like plotters, which will
scale bin values by bin size for density-like combiners, is now applied
by the Normalisation mode.  This keeps the scaling in one place,
and gives the Normalisation mode a chance to ignore the Combiner.Type
if using it would cause trouble (because the same scaling is already
being applied by the Normalisation instance).