ttools: add Unit scaling for histogram-like plots
Where a density-like combination method is used for 1-d histogram-like
plots, a Unit config key is optionally provided.
For a normal 1-d plot this isn't useful, but for a time-like X axis,
it means that you can specify the unit value for density calculations,
letting you choose counts-per-hour, or flux-per-year or whatever.

This replaces the functionality removed a couple of commits back
that was previously handled by different normalisation options.
The advantage of doing it this was is that you can choose
different units for various different combiner modes without
massively inflating the number of normalisation options.
The disadvantage is that you have a selector that is sometimes
doing nothing (if the combiner option is not density-like).