ttools: add sky area unit config to HealpixPlotter
The HealpixPlotter now has the possibility to specify the unit
of sky are when using the density-like combiner modes.

There is some other rewriting too.  The new density-like modes mean
that there's no need to provide a custom Combiner-mode-a-bit-like-mean
that was there before, so the combiner options are now treadted
the same as for SkyDensity (though with somewhat different documentation).