ttools: tweak examples following density plot changes
For the HealpixPlotter, I can now plot per unit of solid angle,
so change the plot to display references per square arcminute,
which is more physically appealing.

For the KDE and KNN, the default combination mode has changed
to sum-per-unit, and so the vertical scaling of the example
plots has changed.  So replace the images.

The colour ramp is now labelled better by default, taking account
of aux axis quantity units.  In one case remove the explicit
labelling of the aux axis, since the default labelling is now
good enough.