ttools: document density-like combiner options
Add a note in the stilts and topcat user document change logs
about the new density-like Combiner options, sum-per-unit and
count-per-unit, update some figures, update some discussions
in the text.

This is not perfect.  The way that the documentation is done
makes it hard to document certain parts of this, e.g. the per-unit
option for histogram-like layers, which only show up in the time plot;
since documenation is by layer, this option never gets documented,
though the information is present in the source code.