ttools: add AuxReader.getAxisInfo method
Per-layer AuxReaders now report their own aux scale coordinate
metadata.  Previously the only way to determine this was to look
at the weight coordinate applying to an AuxScale-bearing layer
and use that metadata.  But sometimes that's misleading, since
the values that form the scale are derived from, but not the
same as, the weight coordinate (e.g. scaled by some axis unit).

Now layers have the chance to report their aux scale metadata
correctly.  This particularly applies to layers that use a Combiner.

Utility methods are also added in PlotUtil to extract suitable
axis metadata from a list of plot layers using this new method.