ttools: histogram-like per-unit specifiers behave better
The various histogram-like plotters now use a PerUnitConfigKey
alongside their ConfigKey<Combiner>, so that the combiner
specifier's GUI can be disabled when a non-unit-sensitive
combiner is selected.  This should make the GUI more comprehensible.

Note it is only in effect for the Time plot (time per-units are
configurable), it makes no difference for the Plane plot.