ttools: add VOSITables-v1.1.xsd
Add new VOSTables v1.1 schema to taplint's set of cached documents,
and fix it so that it is retrieved in response to requests for the
resource http://www.ivoa.net/xml/VOSITables/v1.0 (sic).
The minor version auto-upgrade is to do with the newish XML versioning
policy described in the XMLVers draft PEN.  Added a comment to this
effect in IvoaSchemaResolver to avert of confusion.

The VOSITables-v1.1.xsd document was downloaded directly from
http://www.ivoa.net/xml/VOSITables/v1.0/VOSITables-v1.1.xsd,
or equivalently (via a 303 redirect) from
http://www.ivoa.net/xml/VOSITables/v1.0.

Although VOSI 1.1 is not very new (REC-VOSI-1.1 is dated 2017-05-24),
the v1.1 schema has for one reason and another only just found its
way to its canonical location on the IVOA website.