ttools: avoid multicone queries with sr=NaN
Cone search queries sent using the multi-cone framework in stilts
or topcat were getting sent even if no definite search radius was
provided (SR=NaN).  Such queries are unlikely to give a useful
result, and may be an annoyance to service providers, so this fix
prevents such queries from being sent.

Checks for blank RA and DEC values were already in place.
The reason for the discrepancy is that the same framework is used
for Cone, SIA and SSA queries, and in the latter cases it can make
sense to submit a blank radius.  The check now applied to cone searches
radius is in a slightly different place to the RA/Dec checks, which means
the behaviour is slightly different (blank RA/Dec provokdes a WARNING
message through the logging system, but blank SR does not).
This discrepancy is slightly suboptimal, but not enough of an annoyance
to persuade me to do the necessary reorganisation to avoid it.