votable: preserve aux data in VOTable PARAMs
Modify VOTable PARAM writing so that table parameters (DescribedValues)
with ValueMembers that are in fact ColumnInfos rather than just
(Default)ValueInfos retain their ColumnInfo-specific metadata,
specifically the AuxData items.  As long as this gets passed along,
the PARAM writing code is able to include AuxDatum-like metadata
such as VOStarTable.REF_INFO etc.

Arguably, it shouldn't really do that since AuxData was originally
only supposed to be attached to columns and not parameters.
But in fact there's no good reason for that, and it stops you
doing things you need to do (additional VOTable-specific elements
on output PARAMs).

This is messy; really ValueInfo and not just its ColumnInfo sub-type
should know about AuxData.  But for now this hack improves matters.