topcat: think about better Classification window behaviour
I had a go at improving the ClassifyReportPanel so that it generates
SyntheticRowSubsets rather than non-Synthetic ones.
This would be desirable for STILTS and session serialization.

Didn't succeed, but got as far as realising it's quite hard, though
not impossible.  Made some source code changes and added some
source code comments that should act as a placeholder if I return
to this one day.

No behaviour change.