ttools: remove votlint PARAM->FIELD XML reference warning
XML references from the ref attribute of a PARAM to the ID attribute
of a FIELD in a VOTable document used to provoke an "is this meaningful?"
warning, since although valid according to the VOTable XML schema,
is wasn't obvious that it actually meant anything.

This usage is now however prescribed by the Service Descriptor
construction described in section 4 of the DataLink 1.0 document,
so it is a reasonable thing to encounter in current VOTable documents.
Therefore, the warning is removed, and this usage now passes without
comment.