votable: write multiple tables one per RESOURCE
When outputting multiple TABLE elements in a single VOTable
document, now write each one in its own top-level RESOURCE element.

This is a change in the output format when multiple tables
are written, which is ceteris paribus undesirable, since it's
possible that other software will be confused by this change
of behaviour.  However, it seems to be necessary to group elements
related to different tables together in the same document,
specifically DataLink-style Service Descriptors,
which are encoded as separate RESOURCE elements.
If they all get stored in the same flat top-level RESOURCE element,
there's no good way to determine which service descriptor corresponds
to which TABLE.

For single-table VOTable documents, the output is not changed.