dpac: remove dependencies from commons-math functions
Remove parts of the Apache commons-math functions I've pinched
that are not needed by the parallax code.  Having done this,
no external dependencies are required beyond these three short
source files.  The only important parts that I've taken out
are some checks on the construction time parameters, for instance
that supplied arrays are monotonic.  I'm assuming that the
parallax code is already configured to avoid doing that.

I haven't checked the Apache license to work out what else I
need to do, if anything, to allow me to use and distribute
this code.