dpac: adapt epoch transformation for standalone use
Modify the EpochTransformation and GVector3d classes so that no
external dependencies are required.  This mostly consists of
removing parts of GVector3d that are not required here.

These classes are now in the package uk.ac.starlink.dpac.epoch,
rather than their original location in the gaia.cu1.tools hierarchy.
This is to avoid problems if they are used at some point on
the same classpath as the GaiaTools versions; these classes
are not quite the same, since e.g. they don't implement the
same interfaces.