ttools: add astrometry tests for the Gaia functions
Note that the tests for epoch propagation use Gaia DR2 data, which is
at the time of this commit to the public repository under embargo.
For the purposes of this commit therefore, the data rows have
been removed from the test file.  They will be reinstated to
the repository following DR2.