topcat: document JEL functions in full in SUN/253
The documentation of the various expression language functions
in the user document for TOPCAT has always been less complete
than in the user document for STILTS.  That was historically
because you could find the information in the GUI Available
Functions window in TOPCAT.  But actually you can do that in
STILTS now, and in any case, it's a false economy; it's useful
to have the full information in the user document (which is
also searchable).

So modify the function descriptions in SUN/253 to have the same
level of detail as in SUN/256.  Basically that means including
the parameter and return value descriptions for each function as
well as the textual description of the function behaviour.