ttools: include function javadoc @examples in SUNs
The doclets that turn JEL expression language function javadocs
into XML for inclusion in the TOPCAT and STILTS user documents
SUN/253 and SUN/256 now output the @example entries as well
as the descriptions, parameters and return values.
The output XML formatting is changed a bit as well;
it's now more readable (at least in the HTML on my browser).

The user documents now contain effectively the same information
as is available in the GUI MethodWindow function browser.