ttools: add @HideDoc annotation
Provide a mechanism for hiding some of the functions that are present
in the expression language from the auto-generated documentation.
This may be useful if you want to provide functions that are
available to users at runtime, but whose documentation for whatever
reason would clutter up the auto-generated documentation in an
undesirable way.  Methods that need to be hidden should sport the
@uk.ac.starlink.ttools.build.HideDoc annotation.