ttools: rework Arrays slice/pick functions
Rewrite the user slice and pick functions from a few commits back.
The initial implementation had slice, sliceInt and sliceString
that forced input arrays to a given type and output a fixed type
as per the name.  That is how the other methods in this class
work, but for these it probably makes more sense to have multiple
overloaded methods with return types that match the input type.
I had originally avoided doing that so as not to clutter up
the auto-generated documentation with a lot of same-named
methods which would probably confuse users.  However, now I've
added mechanisms to hide selected methods from the auto-generated
docs, so provide slice/pick in the overloaded form instead.