dpac: add Edsd class and helpers
Add classes for custom evaluation and integration of the
Exponentially Decreasing Space Density distance estimation prior.

This duplicates some of the code in gaia.cu9.tools.parallax,
but extends it in a way that makes it possible to do the
calculations (especially, perform the numerical integration)
more efficiently with similar accuracy.

It also succeeds in PDF evaluation in some regions of parameter
space in which the CU9 code runs into overflows.