ttools: re-implement Gaia distance estimation functions
Rework the distance estimation functions in func.Gaia to use
the Edsd class rather than the DPAC-provided classes.
This is considerably more efficient, because it uses adaptive
quadrature to perform the numerical integration of the PDF
when calculating the quantiles.

Also remove the distance modulus estimation functions.
Although these are provided in the DPAC code, and could be
re-implemented more efficiently in the same way as the distance
estimation functions, they have not been discussed in the
Bailer-Jones papers or elsewhere as far as I know, and looking
at the results their interpretation seems complicated - e.g.
the PDF has only one mode, so picking the lower mode when one
exists doesn't work.  This tends to dramatically overestimate
the distance modulus in some cases even when that doesn't
happen for the distance estimate.