ttools: document comment syntax in filter (cmd) parameters
Document the fact that in filter commands using file indirection
(e.g. cmd=@script.txt) you can use comment lines or blank lines.
In fact this is true even for non-indirected values, but not
very useful in that case.

This has always been the case, but the fact it was undocumented
led to me re-implementing it before I realised it was already
there, so maybe the documentation will stop that happening next time.