topcat: adjust ClassifyWindow subset name assignment
Adjust the way that subsets auto-generated by the Column Classification
Window are named by default.  The names are derived from the values
truncated to a certain length.  That length was 6 characters;
increase it to 16.  It's probably still not too long to be
unwieldy, but gives a better chance of having unique values
for all subsets.  Suggested by Peter Draper.