topcat: limit width of table selector in relevant windows
Use the modified TablesListComboBox to make sure that table
selectors in topcat windows do not become indefinitely wide
in the presence of tables with very long labels.
This is a bit smarter than just setting a maximum width,
it is sensitive to the contents of the ComboBoxModel to
work out how wide the JComboBox should actually be,
but it doesn't go wider than a given limit.

This is done everywhere the table selector appears, except
for an obsolite legacy class in (topcat.plot.PointSelector).