topcat: add JEL-related ActivationTypes
New ActivationTypes JelActivationType and ShellActivationType.
The JEL one corresponds to the old "Execute Custom Code" option,
but provides a less cramped GUI and better feedback on results.
The Shell one provides a more convenient way to execute OS
commands; this could be done using the JEL activator with
the System.exec() functions, but this provides a nicer GUI.