topcat: add classes for working with DataLink tables
Various functionality related to DataLink {links}-response tables
in TOPCAT.  This includes configurable invocation of URLs,
and a new ActivationType to view a URL as a Datalink table.
A DatalinkPanel which displays a links table and UrlPanel which
offers various ways to invoke URLs selected from it are included.

There is also a framework for defining different resource types,
the various things you might want to do with URLs for said types,
and preferences for those options.