topcat: add new activation framework
Set up a new ActivationWindow in TOPCAT and a set of interfaces
that can be implemented to provide different activation actions.
The basic idea, of getting something to happen when a row is
clicked on or otherwise selected, is the same, but there are
many improvements, and the GUI is completely changed.

The old ActivationQueryWindow is replaced by a new ActivationWindow.
This is a table view window, alongside the other ones like columns
and parameters, rather than a special window from a custom button.
This allows you to choose multiple activation actions simultaneously,
and offers more information about what's available and what's useful.
Feedback from the activation results is displayed in the window.
It's more extensible and it's pluggable.

At this commit, there is no documentation, and no implementations
of the new ActivationType interface apart from one dummy one.