topcat: fix SAMP activators to send URLs not filenames
The various Send*ActivationTypes are documented to accept "location"
strings that can be either a filename or a URL.  However, they were
sending this location directly as the "url" argument of their SAMP
messages.  In the case that filenames were used, this meant that
receiving clients (usually) don't recognise them, since they're
not URLs.

Add code that makes sure filenames are turned into URLs before
sending them in SAMP messages.