ttools: encourage user to delete unwanted TAP async jobs
When tapquery is run in async mode and the user gets bored and
hits ctrl-C, the job is left running.  This probably happens
quite a lot, leading to wasted CPU cycles running jobs that
nobody ever sees the results of.

This commit fixes it so that if a tapquery job is still running
at JVM exit, a message is written on stderr informing the user
of the fact, and suggesting they might like to delete it.

A new deletion mode, "delete=now" is added to the tapresume task
to make this easier, and the corresponding tapresume invocation
is suggested.  Tapquery also has delete=now, though it's not
very useful in that case.