topcat: add more items to ActivationWindow toolbar
Add toolbar items for the existing Add and Remove actions
(previously only in the menu).  Add is a bit clunky, since
it requires a popup menu from the toolbar, but it's important
to have it on display.

Also add a new RemoveInactive action to toolbar and menu.
Add remove-all-inactive-entries button to toolbar and menu