ttools: adjust JSON library to retain Object key sequence
Make a small custom change to the json.jar library;
I have just replaced the use of HashMap by LinkedHashMap passim
in JSONObject class.  Other than that, this is built from source
from the a9a0762 commit that was the previous source of this library
(local commit is f033da0).

This shouldn't have any effect except making sure that when serializing
a JSONObject created from a Map, the entries are written in a
predictable order, namely the iteration order of the originating Map.
JSON objects are supposed to be unordered, but this can provide
a cosmetic improvement that's useful during debugging, diagnosis,
or generally looking at serializer output.