util: fix HeadBufferInputStream integer overflow bug (TAP)
HeadBufferInputStream was using an int to keep track of
stream position instead of a long.  This class is (only) used
by uk.ac.starlink.vo.TapQuery when reading result VOTables.
The effect was that if a VOTable result of a TAP query (sync or async)
was longer than 2^31 bytes, an java.lang.ArrayIndexOutOfBoundsException
was reported with an index near to -2^31.  Now uses a long, so should
be OK.