ttools: make fallback RangeSlider a bit less ugly
Change some of the UI details for the RangeSlider that's used
when the JIDE one is unavailable.

The colours are now taken from the UIManager rather than being
lurid red/green/blue as before.  It also just inherits the
track drawing from its superclass.  This is slightly less
visually informative, since the track between the thumbs is
no longer distinguished from that outside them, but users
can work out what's going on, and the visual effect before
was too clashing.

I experimented with inheriting from the slightly more
aesthetically pleasing plaf.metal.MetalSliderUI rather than
plaf.basic.BasicSliderUI, but that results in a NullPointerException
on OSX Java 10 (though I didn't get to the bottom of exactly
what's causing it), so don't do that again without careful testing.

It's still not beautiful, but it's better.