topcat: use java.awt.Desktop not BrowserLauncher2
Java 6 introduced the method java.awt.Desktop.browse(URI) which
lets you hand off a URL to a desktop system web browser.
This does the same job as the external BrowserLauncher2 library
I'd been using for a long time - presumably because the Desktop
class wasn't in the J2SE when I first wanted to do that.

I had at least one (probable) report of BrowserLauncher2 not
working, which is not very surprising, since it contains all
sorts of OS-specific code and was last updated in 2007.

So remove all uses of BrowserLauncher2, and replace them by
uses of Desktop.browse.  It still may not be bulletproof,
but it's more likely to work than the third-party code.