topcat: fix activation-related headless bug in unit tests
A bug was introduced into the unit tests 8 commits ago (0d90808)
which meant that the test uk.ac.starlink.topcat.CodecTest failed
with a java.awt.HeadlessException when run headless
(-Djava.awt.headless=true).

This was because the Codec now calls TopcatModel.getActivationWindow,
since the activation window is where activation state is stored.
This is not very good practice, the activation state should really
be in some class that does not contain any display components
(MVC, or at least MV).  However, it's difficult to extricate the
state from the UI, so till now I haven't done it that way.

Hack round this by creating the ActivationWindow lazily and
making sure that no such window is created when unit tests are run.

It would be better to eliminate the windows from the activation state,
and similarly with the ControlWindow, which also contains state.
Maybe one day.