topcat: add import params option for load/plot table action
The {Load,Plot}TableActivationType configurators now have a
new option, import table parameters.  It adds all the values
in the activated row as parameters to the imported table.
This means you have access to the properties of the activated
row (using the param$ syntax) when the activated table is imported.