topcat: set OSX Info.plist JVMVersion to 1.8+
Following discussions with Timo Prusti, I have changed the
JVMVersion attribute setting Info.plist file that configures
the DMG file built for MacOS to "1.8+" (the default was "1.3+").
Tests indicate this probably won't make a difference, but it
might cause the runtime to run it under 1.8 (which avoids at
least a bug relating to scrambled JTables).

This might possibly cause trouble to people who only have java 1.6
installed on their Mac.  If so, that's not necessarily bad, since
I may move to a 1.8 target soon, so it would be useful to find
out about such issues now.