topcat: move Activation Functions manual section
The Activation Functions section of SUN/253 (id="topcat.funcs")
has been withdrawn, and the content moved into the Execute Code
activation action documentation section, which is now more
appropriate.  Also add a note that some of these items may be
withdrawn in future (they are quite out of date and in many cases
not much use).