ttools: slight updates to jdbcConfig manual sections
Expanded the JDBC Configuration section of SUN/256 to include more
detail on different drivers (as in SUN/253), at the request of
Marco Molinaro.

At the same time updated the content of this section in
both SUN/256 and SUN/253 to update/remove several broken links
to external RDMBS corporate web pages, and add a note that
the text is known to be incomplete and out of date.