ttools: provide skygrid lon and lat grid crowding separately
Replace the SkyGridPlotter crowd parameter with two parameters,
loncrowd and latcrowd, so that density of parallels and meridians
can be controlled separately (e.g. to plot just an equator and no
other grid lines).  This was requested by Sergey Klioner (DPAC/AGIS).

This is a backwardly-incompatible change: invocations which use the
crowd parameter for this plot layer will now fail.  It would be
possible to catch that parameter value to prevent that happening,
but probably it's better to advertise that the new options are
present.  I don't expect the impact will be very high, since
probably not many people are using this plot layer.