ttools: adjust auto-ranging near zero
When auto-ranging data bounds for a plot, if one of the range
limits is near zero, until now it has been set to exactly zero
(rather than applying the usual padding factor) since that's
usually what the user would want to see.  However, that can
result in values right on the axis=0 line being excluded from
the plot, even if they were used to set the auto-range values.
This effect is seen at the right and bottom edges (increasing graphics
X and Y coordinates), and can result in valid data being hidden
in the default plot configuration, which is quite serious.

So fix it to make sure there is always at least a tiny amount of
padding to avoid this problem.

Unfortunately this introduces pixel-level artefacts in some cases:
e.g. Y axis zero value moved up by a pixel, and a pixel-height gap
at the bottom of KDE plots etc.  These are not wrong, just aesthetically
suboptimal.  The user can avoid them by setting the X/Y range explicitly,
i.e. bypassing or overriding the auto-range.

Other options to ameliorate this are conceivable, for instance
adjusting the way that data coordinates are mapped to graphics
coordinates, by use of rounding instead of truncation, or forcing
truncation always to operate in the direction of decreasing data
coordinate.  But such things would have wide ranging consequences
in the plotting framework and would be fraught with possibilities
for hard-to-spot problems.