ttools: follow HTTP 3xx status in more cases
Fix tools votcopy, votlint, datalinklint (possibly others?) so that
they follow HTTP 3xx redirects in most cases when fed an HTTP URL
as input.  The table pipeline commands already do this.
In some cases the direct input parameters did not do that processing,
and in some cases they did it by using TapQuery.followRedirects,
which follows only 303s (as per the TAP specification), but not
statuses 301, 302, 307.