ttools: ensure proper SliderSpecifier monitor text display
I had a report that sometimes the text in the slider+text text
box is displayed with the right hand end (least significant part)
of the number visible rather than the left hand end.
That shouldn't happen, since the numeric value ought to be
formatted in such a way that it stays short so the whole thing
is visible, and I can't reproduce it, but add a fix for the
case in which it's a long string representation in any case.

Also add a logging line to repor the error I think might be
causing this to happen.