topcat: improve column search window
Reimplement some of the GUI interaction in the recently introduced
ColumnSearchWindow.  It's not as complicated as I thought.

It now operates faster, and row activation only occurs if exactly
one row is located (in the previous implementation it occurred
if any number >0 of rows was found, applying to the first one
encountered, which was probably unwanted behaviour).