topcat: fix i18 slider text bug
SliderSpecifier uses PlotUtil.formatNumber to format numbers to a
limited precision and then re-parse the result to give a round
number for reporting in the text widget.  Up till now formatNumber
was returning an internationalised stringification of the number;
in e.g. -Duser.language=fr it used a comma instead of a period for
decimal point, which messed things up.  Fix it to force use of
Locale.UK so that doesn't happen.