topcat: load progress indicators format long row counts
When loading tables, TOPCAT reports the number of rows loaded
so far in a couple of places.  This can be hard to read if the
number gets very large.  So use NumberFormat to format them
(in a Locale-sensitive way), e.g. adding commas every three digits
to increase readability.

I don't think that this (implicit in the source code) Locale
sensitivity is going to be problematic, but if it is, be prepared
to force to a particular Locale.