vo: adjust exclusion of TAP_SCHEMA.columns size column
The size ("size") column in TAP_SCHEMA.columns is a pain, because
it's a reserved word.  So it's avoided when interrogating the
TAP_SCHEMA tables for metadata.  But this was also throwing out
the (harmless and useful) arraysize column introduced at TAP 1.1.
Fix it so that arraysize is allowed through, and be more explicit
in the comments about what's going on.