vo: improve TAP column metadata display
Make a few adjustments to the JTable that's displayed in
the Columns tab of the TAP metadata browser.

The main thing is to replace the (low-level) DataType column with
the more user-friendly Type column; this contains text that describes
both the datatype and the (newly available) arraysize.
Previously, the arraysize wasn't visible.

The TAP 1.0 "size" column still isn't visible, because it's a pain
to handle.  Arguably it should be, but since nobody has complained
about it up till now, and TAP 1.1 is now being implemented, I'm not
going to attempt it at least until somebody asks me to.

A couple of more cosmetic things: reorder the columns a bit, and
also include the new (TAP 1.1) Xtype item.