table: adjust system save progress popup text
The progress popup window you get when saving a table
having chosen the location using the System Browser option
was reporting a message like "Saving table [name] to [destination]".
But the [name] is not always a recognisable value, and a user
reported that it can be confusing.  So just report
"Saving table to [destination]" instead; the other information
isn't really that necessary or useful.

This was pointed out by J-B Marquette; add his name to the topcat
acknowledgements list for this and quite a few other smallish
items over the years.