topcat: add Activation Action security blocking mechanism
Add capabilities to the Activation Window to work with activation
actions loaded from sessions that may be insecure, i.e. potentially
the target of injection attacks.  In principle a bad person could
prepare a TOPCAT session file and persuade a third party user
to load it, invoking arbitrary code with the user's privileges.

To address this, work out which activation actions loaded from
a session file could conceivably be malicious, and require the
user to approve them explicitly before they are ever invoked.
Actions not loaded from a session file are not treated in this way,
since the user must have prepared them him/herself, so should be
aware of any potentially dangerous implications.

This is not documented in SUN/253 yet.