util: add allowSystem param to DataSource cmdline methods
Some of the DataSource methods have a location parameter that is
allowed to include a shell command pipeline to produce the
input stream.  This is a potential security risk in the case that
the location string may come from somewhere vulnerable to injection.

This commit adds a boolean allowSystem parameter to those methods,
so that calling code can decide whether to allow these system
commands.  If false, no system command evaluation is performed.

Client code is updated so that task parameters have allowSystem=true,
since these are presumably set directly by the user, but for values
that may come from table values or saved session information
(this mainly means activation actions) allowSystem is set false.
This ought to prevent bad persons injecting any nasty surprises.