topcat: allow safety configuration in some activation types
For activation actions ViewImage, LoadTable and PlotTable,
provide configuration options within the Configurator GUI
so that the user can control whether unsafe (system command
preprocessing in data source location) location access will
be permitted.  By default these are switched off, so that these
actions normally will not be subject to security-related blocking,
but if the user wants to use that facility, they now have the
option to turn it on.