topcat: document activation security features
New section activationSecurity added to SUN/253, with links from
those activation actions that might trigger it.

Is it a good idea to include this information in the user document?
You could argue (security through obscurity) that it might give
people ideas they might not otherwise get.  On the other hand,
nobody reads the documentation, so it probably doesn't matter.