ttools: add aux shading to Line3dPlotter
Line3dPlotter now has a Aux coordinate that can be used to change
the colour of the line along its length.

This is a bit anomalous in a few ways.  The aux shading should
probably be an option, available (at least apparently in the UI)
in the same way as marker shading modes, though currently only
Aux and Flat will be possible.  That might be possible by
implementing the ModePlotter marker interfaces, but I haven't
done it yet.

Also, this aux shading does not exist for the 2d line plotter.
Providing it for 2d would require some implementation changes there.
However, it's not as necessary for 2d (though it would sometimes
be nice), since 2d line plots tend not to be so confusing.