ttools: adjust contour level calculations
The contour levels were not getting calculated well in most cases
for low point densities.  That's because the levels were being
generated from the unsmoothed pixel density grid rather than the
smoothed grid in most cases, which is in turn because the smoothing
was done using a separable (Gaussian) smoothing kernel in 2 1-d
steps rather than 1 2-d step; that's faster but has a hard time
distinguishing between zero and missing values.
I was trying to work round that, but it turned out not to be very well,
so I adjusted the separable kernel smoothing so it calculates a
mask as well and can therefore distinguish between zero and missing
data in the smoothed grid.  That all means that the contour levels
work OK for low density grids.

That in turn made the default contour plots kind of ugly because the
levels tended to take in isolated points, so I adjusted the semantics
and default of the OFFSET (zero point) parameter as well - it's no
longer folded to the range 0..1, and it defaults to 1 rather than 0,
so that the first level is typically away from the zero value of
combination/density.

At the same time add a numeric view of the zero point value
alonside the slider.

The original issue was pointed out by Bertrand Goldman.