ttools: new class SphereAngleRange
Add a class that can characterise and calculate the range of theta,phi
space occupied by a given cube.

This was written in preparation for use with the sphere grid plotting,
but it may not be used for that.  Note that it has received no testing,
and may not be used anywhere.  It may come in useful though.