ttools: change serialization of Gaia test VOTable
Test file ap2b.vot in the func.Gaia unit tests was committed in
BINARY2 format, I've changed it to TABLEDATA.  This is less
compact, but it means that I can make changes to it with it
(a) being obvious from repository diffs what they are and
(b) requiring smaller commit blobs.

Conversion was done using stilts votcopy.