topcat: modify TopcatJELRowReader implementation
Change the implementation and constructor signature of TopcatJELRowReader.

Now, instead of supplying a fixed list of all RowSubsets at
construction time, the RowReader identifies any subsets that are
needed by expressions it's asked to compile, and stores those
internally.  This makes for more straightforward construction
(just pass a TopcatModel) and it also means that it doesn't hang
on to references for unused RowSubsets that may otherwise become
eligible for garbage collection if those subsets are removed
from the TopcatModel.

The behaviour should be unchanged; expressions involving deceased
RowSubsets can still be evaluated.