topcat: work round JavaHelp indexing bug
The previous-but-one commit (e0f3f0783f) introduced a problem with
the JavaHelp index file creation that happens as part of the
build process.  During execution of the docs-html target, it
started emitting the following message:

 [java] I/O exception occurred in file 'versions.html'

and I think it prevented completion of the help indexing.

This message was emitted by the invocation in the build file of the
class com.sun.java.help.search.Indexer.  It's the result of an
IOException which is swallowed by the code, but following a bit of
detective work the relevant stack trace seems to be this:

    java.io.IOException: Can't store title
        at com.sun.java.help.search.HTMLIndexerKit.parse(HTMLIndexerKit.java:133)
        at com.sun.java.help.search.Indexer.parseFile(Indexer.java:254)
        at com.sun.java.help.search.Indexer.parseFile(Indexer.java:266)
        at com.sun.java.help.search.Indexer.compile(Indexer.java:214)
        at com.sun.java.help.search.Indexer.main(Indexer.java:86)
    Caused by: java.lang.ArrayIndexOutOfBoundsException: 256
        at com.sun.java.help.search.BitBuffer.concatenate(BitBuffer.java:166)
        at com.sun.java.help.search.Compressor.concatenate(Compressor.java:58)
        at com.sun.java.help.search.DocumentCompressor.encodePositions(DocumentCompressor.java:234)
        at com.sun.java.help.search.DocumentCompressor.encode(DocumentCompressor.java:198)
        at com.sun.java.help.search.DocumentCompressor.compress(DocumentCompressor.java:118)
        at com.sun.java.help.search.DefaultIndexBuilder.closeDocument(DefaultIndexBuilder.java:147)
        at com.sun.java.help.search.DefaultIndexerKit.endStoreDocument(DefaultIndexerKit.java:154)
        at com.sun.java.help.search.HTMLIndexerKit.parse(HTMLIndexerKit.java:131)
        ... 4 more

This appears to be the same issue noted at:

   https://github.com/javaee/javahelp/issues/38

reported originally in 2012 and imported to github in 2017 from
java.net JIRA issue JAVAHELP-38.  Sounds like it's not straightforward
to fix.

To cut a long story short, it looks like the problem in this case
was triggered by one too many occurrences of the word "add" in
the file that ends up as versions.html in the Help library
(the version history section at the end of SUN/253).

After playing around with the invocation flags reported by running
Indexer -h, it seems that adding the flag "-nostopwords" to the
Indexer invocation fixes the problem in this case.
I thought I was going to have to add the term "add" to the stop words
list, but it doesn't seem to be necessary.  But if this happens again
for a different word, maybe that will be necessary.